src/test/java/homework6methods/FirstMethodTest.java
package HomeWorkMethods;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.util.Arrays;

public class FirstMethodTest {

    @Test
    public void getNewArrayAfterLastFourFromArrayFirstTest() {
        int[] intArray = {1, 2, 3, 4, 5, 4, 1, 7};
        int[] result = {1, 7};
        Assertions.assertArrayEquals(result, FirstMethod.getNewArrayAfterLastFourFromArray(intArray), "Test failed");
    }

    @Test
    public void getNewArrayAfterLastFourFromArraySecondTest() {
        int[] intArray = { 6, 7, 8, 9, 10};
        try {
            FirstMethod.getNewArrayAfterLastFourFromArray(intArray);
            Assertions.fail("Test failed");
        } catch (Exception ignored) {

        }
    }

    @Test
    public void getNewArrayAfterLastFourFromArrayThirdTest() {
        int[] intArray = {4};
        FirstMethod.getNewArrayAfterLastFourFromArray(intArray);
        Assertions.assertTrue(FirstMethod.getNewArrayAfterLastFourFromArray(intArray).length == 0);
    }
}
